+++ Core/RA.Utilities.Core.Exceptions/BadRequestException.cs
@@ -12,15 +12,24 @@ public class BadRequestException : RaBaseException
     /// Initializes a new instance of the <see cref="BadRequestException"/> class with a single validation error.
     /// </summary>
     /// <param name="error">The validation error that caused the exception.</param>
-    /// <param name="errorCode">A specific error code associated with the error.</param>
-    /// <param name="responseCode">The error code. Defaults to 400 (Bad Request).</param>
-    /// <param name="message">The error message. Defaults to a standard bad request message.</param>
     public BadRequestException(
-        ValidationErrors error,
-        string errorCode = nameof(BaseResponseCode.BadRequest),
-        string message = BaseResponseMessages.BadRequest,
-        int responseCode = BaseResponseCode.BadRequest
-    ) : base(errorCode, message, responseCode)
+        ValidationError error
+    ) : base(BaseResponseCode.BadRequest, ResponseType.BadRequest, BaseResponseMessages.BadRequest)
+    {
+        Errors = [error];
+    }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="BadRequestException"/> class with a single validation error.
+    /// </summary>
+    /// <param name="error">The validation error that caused the exception.</param>
+    /// <param name="errorCode">A specific error code associated with the error.</param>
+    /// <param name="message">The error message.</param>
+    public BadRequestException(
+        ValidationError error,
+        int errorCode,
+        string message = BaseResponseMessages.BadRequest
+    ) : base(errorCode, ResponseType.BadRequest, message)
     {
         Errors = [error];
     }
@@ -29,57 +38,75 @@ public class BadRequestException : RaBaseException
     /// Initializes a new instance of the <see cref="BadRequestException"/> class with a collection of validation errors.
     /// </summary>
     /// <param name="errors">An array of validation errors that caused the exception.</param>
-    /// <param name="errorCode">A specific error code associated with the error.</param>
-    /// <param name="responseCode">The error code. Defaults to 400 (Bad Request).</param>
-    /// <param name="message">The error message. Defaults to a standard bad request message.</param>
     public BadRequestException(
-        ValidationErrors[] errors,
-        string errorCode = nameof(BaseResponseCode.BadRequest),
-        string message = BaseResponseMessages.BadRequest,
-        int responseCode = BaseResponseCode.BadRequest
-    ) : base(errorCode, message, responseCode)
+        ValidationError[] errors
+    ) : base(BaseResponseCode.BadRequest, ResponseType.BadRequest, BaseResponseMessages.BadRequest)
     {
         Errors = errors;
     }
 
     /// <summary>
-    /// Gets or sets the list of validation errors.
+    /// Initializes a new instance of the <see cref="BadRequestException"/> class with a collection of validation errors.
     /// </summary>
-    public ValidationErrors[] Errors { get; }
+    /// <param name="errors">An array of validation errors that caused the exception.</param>
+    /// <param name="errorCode">A specific error code associated with the error.</param>
+    /// <param name="message">The error message.</param>
+    public BadRequestException(
+        ValidationError[] errors,
+        int errorCode,
+        string message = BaseResponseMessages.BadRequest
+    ) : base(errorCode, ResponseType.BadRequest, message)
+    {
+        Errors = errors;
+    }
+
+    /// <summary>
+    /// The list of validation errors.
+    /// </summary>
+    public ValidationError[] Errors { get; }
 }
 
 /// <summary>
 /// Represents a single validation error, providing detailed context about a failure.
 /// </summary>
-public class ValidationErrors
+public class ValidationError
 {
     /// <summary>
-	/// Gets the name of the property that failed validation.
-	/// </summary>
-	[JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string PropertyName { get; init; }
+    /// Initializes a new instance of the <see cref="ValidationError"/> class.
+    /// </summary>
+    /// <param name="errorMessage">The message describing the validation error.</param>
+    public ValidationError(string errorMessage)
+    {
+        ErrorMessage = errorMessage;
+    }
 
     /// <summary>
-    /// Gets the message describing the validation error.
+    /// The name of the property that failed validation.
+    /// </summary>
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public string? PropertyName { get; init; }
+
+    /// <summary>
+    /// The message describing the validation error.
     /// </summary>
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public string ErrorMessage { get; init; }
 
     /// <summary>
-    /// Gets the value that was provided for the property, which caused the validation to fail.
+    /// The value that was provided for the property, which caused the validation to fail.
     /// </summary>
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public object AttemptedValue { get; init; }
+    public object? AttemptedValue { get; init; }
 
     /// <summary>
     /// Gets a custom error code associated with the validation failure.
     /// </summary>
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string ErrorCode { get; init; }
+    public string? ErrorCode { get; init; }
 
     /// <summary>
-    /// Gets or sets the expected value.
+    /// Gets the expected value, if applicable.
     /// </summary>
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public object ExpectedValue { get; set; }
+    public object? ExpectedValue { get; init; }
 }